repl: fix strict-tooling violations across REPL TUs
Phase 1 strict gates (test-tidy-strict + test-cppcheck +
test-docstring-coverage) caught a set of issues against the v0.9.1 REPL
sources:

- POSIX feature macros: declare _POSIX_C_SOURCE=200809L in
  urepl_dispatch.c + urepl_listener.c so clock_gettime / nanosleep /
  CLOCK_MONOTONIC are visible to clang-tidy strict mode.
- Implicit-widening: cap macros now cast to size_t before the
  multiplication (BT_DEFAULT_CAP, UREPL_DEFAULT_OUTPUT_CAP) and the
  1ms-nanosleep literal uses (long) explicit type.
- Multi-level pointer conversions: explicit (void *) cast on realloc /
  free of UJsonNode **; sizeof(UJsonNode *) instead of sizeof(*ni).
- memcpy not-null-terminated false positives: NOLINTNEXTLINE annotations
  on append_lit (urepl_ndjson.c) and the JSON-escape helper
  (urepl_introspect.c) — both append raw bytes; caller manages NUL.
- const-correctness: urbi_introspect_profile/gc/stack now take const
  UVM*; dispatch_cancel_stub/lobby_new/lobby_close take const UReplJob*;
  reader_parse_lines takes const char*; UReplAuthLimiterSlot* / sym /
  target / e locals declared const where read-only.
- variableScope: emit_json_escape's `need` and emit_object_slots's `v`
  moved into the inner-most scope that uses them.
- knownConditionTrueFalse: removed redundant `if (session != NULL)`
  guard in spawn_reader cleanup (session was non-NULL by the urepl_
  session_create success path above).
- urepl_queue.h: docstrings for the 5 public functions flagged by
  test-docstring-coverage (init/destroy/count, ringbuf_init/destroy).
- .cppcheck.suppressions: urepl.c file-wide constParameterPointer
  suppression for urbi_repl_serve_step (signature pinned by
  include/urbi/repl.h).

+++ src/repl/urepl_buffer_transport.c
@@ -23,7 +23,7 @@
 #include <stdlib.h>
 #include <string.h>
 
-#define BT_DEFAULT_CAP (64u * 1024u)
+#define BT_DEFAULT_CAP ((size_t)64u * 1024u)
 
 struct UBufferTransportState {
     UReplRingbuf c2s;    /* client → server */
